Fixing GE-Proton instal for new releases
This break makes sure program won't try to look for another archive
after it finds one for exact version. Even then the way this function
operates and how it tries to correleate archive with checksum file is
flawed as in case host of eg. GE-Proton decides to upload tar.zst first
instead of tar.gz the list of releases will pair tar.zst with wrong
checksum value. If checksum validation is there to say and be future
proof there should be different method for it. This break is more of a
bandaid that will fix it for now